+++ R/zzz.R
@@ -1,2 +1,5 @@
 #' @import rlang
 NULL
+
+# https://github.com/r-lib/tidyselect/issues/201
+utils::globalVariables("where")