+++ tests/testthat/test-preproc_data.R
@@ -1,3 +1,5 @@
+withr::local_package("preproc.iquizoo")
+
 test_that("Basic situation in `preproc_data()`", {
   data <- tibble::tibble(
     user_id = 1:2,
@@ -6,49 +8,28 @@ test_that("Basic situation in `preproc_data()`", {
       jsonlite::toJSON(data.frame(nhit = 1, feedback = 1))
     )
   )
-  skip_if_not_installed("preproc.iquizoo", "1.3.0")
-  library(preproc.iquizoo)
-  dm_indices <- data |>
+  dm_results <- data |>
     wrangle_data() |>
     preproc_data(.fn = bart)
-  expect_snapshot(dm_indices)
-  expect_snapshot(dm::dm_get_tables(dm_indices))
-})
-
-test_that("Can deal with abnormals:", {
-  expect_warning(be_null <- preproc_data(dm::dm()), class = "data_empty")
-  expect_null(be_null)
-  data <- tibble::tibble(
-    user_id = 1,
-    game_data = jsonlite::toJSON(data.frame(nhit = 1))
-  )
-  {
-    be_null <- data |>
-      wrangle_data() |>
-      preproc_data(.fn = bart)
-  } |>
-    expect_warning(class = "data_invalid") |>
-    expect_warning(class = "indices_empty")
-  expect_null(be_null)
+  expect_snapshot_output(dm_results)
+  expect_snapshot_output(dm::dm_get_tables(dm_results))
 })
 
 test_that("Complex dplyr verbs in `preproc_data()`", {
-  set.seed(1)
+  withr::local_seed(1)
   data <- tibble::tibble(
     user_id = 1,
     game_data = jsonlite::toJSON(
       data.frame(
         acc = sample(c(0, 1), 20, replace = TRUE),
-        type = sample(rep(c("left", "right"), 10)),
+        type = sample(rep(c("target", "nontarget"), 10)),
         rt = runif(20, 300, 2000)
       )
     )
   )
-  skip_if_not_installed("preproc.iquizoo", "1.3.0")
-  library(preproc.iquizoo)
-  dm_indices <- data |>
+  dm_results <- data |>
     wrangle_data() |>
     preproc_data(.fn = cpt)
-  expect_snapshot(dm_indices)
-  expect_snapshot(dm::dm_get_tables(dm_indices))
+  expect_snapshot_output(dm_results)
+  expect_snapshot_output(dm::dm_get_tables(dm_results))
 })